Add product management to admin panel
Introduces ProductController and related views for CRUD operations on products. Updates migration to use integer for base_price, adds product routes, and integrates product management into the admin sidebar. Also fixes minor layout issues in bonus withdrawal view.

+++ database/migrations/2025_07_14_145320_create_products_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('products', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->decimal('base_price', 15, 2);
+            $table->integer('base_price');
             $table->text('description')->nullable();
             $table->timestamps();
         });
